refactor(engine): split engine.rs into feature modules
