+++ crates/servo-fetch/src/lib.rs
@@ -13,8 +13,8 @@ pub mod sanitize;
 
 pub(crate) mod bridge;
 pub(crate) mod crawl;
-pub(crate) mod engine;
 pub(crate) mod error;
+pub(crate) mod fetch;
 pub(crate) mod layout;
 pub(crate) mod map;
 pub(crate) mod net;
@@ -25,9 +25,13 @@ pub(crate) mod scope;
 pub(crate) mod screenshot;
 pub(crate) mod sys;
 
-pub use engine::{
-    ConsoleLevel, ConsoleMessage, CrawlError, CrawlOptions, CrawlPage, CrawlResult, FetchOptions, MapOptions,
-    MappedUrl, Page, crawl, crawl_each, extract_json, fetch, init, map, markdown, text, validate_url,
-};
+pub use crawl::{CrawlError, CrawlOptions, CrawlPage, CrawlResult, crawl, crawl_each};
 pub use error::{Error, Result};
-pub use net::NetworkPolicy;
+pub use fetch::{ConsoleLevel, ConsoleMessage, FetchOptions, Page, extract_json, fetch, markdown, text};
+pub use map::{MapOptions, MappedUrl, map};
+pub use net::{NetworkPolicy, validate_url};
+
+/// Set the network policy. Must be called at most once, before any engine use.
+pub fn init(policy: NetworkPolicy) {
+    bridge::set_engine_policy(policy);
+}
